Reduce the number of request necessary to load multiple nodes from the server
Operations such as INode.getChildren can load all nodes at the same time
instead of sequentially.

A new BulkQuery class allows to collect request for these bulk queries
and execute them in one HTTP request, even if they require to load
multiple entries transitively. The number of HTTP request only depends
on the maximum number of request the would be required for one node
and on the number of nodes.