It was possible that property values where recognized as hash keys
The model server detects invalid data by searching for sha hashes in
the values of the submitted entries and checking if the keys exist in
the database.
These sha keys are base64 encoded. Property values are also base64
encoded allowing them to match the regexp for extracting the keys.

Now an additional "*" is added to the sha keys which is a character
that doesn't appear in the base64 encoding output.