Extracted model API to a separate library
The goal is to use the same API for shadow models, which is part of
MPS-extensions. Then the shadow models transformation engine can
directly operate on models stored on the model server.

The modelix model API was orginally copied from shadow models.
Currently, it's very ugly and confusing to integrate two very similar
APIs.