ui-server: allow cloning repositories with authentication
Two new environment variables are used:

GIT_USER: the user name used to clone the repository
GIT_PASSWORD_FILE: path to the file that contains the password.

Password files can be backed by a secret within kubernetes e.g.

```
env:
  - name: GIT_PASSWORD_FILE
    value: "/etc/git-password"
volumeMounts:
  - name: git-password
    mountPath: /etc/git-password
    readOnly: true
volumes:
  - name: git-password
    secret:
      secretName: git-password
```

The part of the system that starts the ui-server is responsible for supplying the
values when the container is created.
If no user or password is provided the repository is cloned without authentication.