ui: add ENV var for heap ratio
New environment variable in the Dockerfile-ui called HEAP_RATIO which controls the max heap ratio of the JVM.
85% is the default but for projects with lots of plugins loaded into MPS memory required for classloading is large.
If lots of memory is allocated for classes the 85% ratio can be to much and the JVM exceeds the memory limit of the container.
Containers build on top of the modelix container can simply override the ENV and increase the value.