Replacement for the GarbageFilteringStore
While modifying CLTree`s a lot of intermediate KV-store entries
are created that are not never used again after a transaction
finishes and a version is created.
The purpose of the GarbageFilteringStore is to buffer all writes
and only write those entries that are reachable from a version
written to a branch.
But the GarbageFilteringStore has no information about which
entries are still needed and which can be removed.
These intermediate entries were never removed from the memory and this
memory leak was a problem during a bulk import.
Reimplemented it in a way that allows the JVM garbage
collector to make this decicion.