Introduced CLHamtSingle, a node that replaces multiple levels of CLHamtInternal with single children
Prevously, the least significant bits of a key where used for the root level. Now they are used for the bottom level. This resulted in small trees it the leafs with long branchless paths to the root. A lot of request are required for these long path. CLHamtSingle makes the paths shorter and therefor reduces the number of requests.