BulkQuery couldn't handle very large data structures
The idea of BulkQuery was to traverse a data structure with the minimum
number of request. For large data structures the number of entries
requested at the same time was to big. HTTP requests failed because one
request was to big and also the pending request queue got to big to fit
into memory.