Added inspector svg editor
Now in the svg view per default the inspector editor is displayed under the main editor.

Because the same socket in the client needs to be shared by two different image based editors we extracted
all the code responsible for managing the connection into a new "SVGContainer" class. The responsibility
of this class is to start and manage the websocket connection and shared it with all the svg image editors
under this same container.