ui-server: allow loading additional plugins
Added a new environment variable that allows loading additional plugins.
This is required if modules used in the the image have a idea plugin
facet. If the idea plugin isn't loaded the plugin facet is marked as
invalid and the complete module is marked as invalid by the class loader.
If this happens all dependent modules are also marked as invalid.

The variable supports a comma separated list of plugin id and optionally
the plugin folder if the folder is different from the plugin id.