Removed mouse move delay. Fixed it on the server side instead.
A delay for mouse move events was introduced on the client side, because
the server couldn't handle that many events. The reason was, that after
each message the editor state was persisted to the model server.
This doesn't happend anymore for mouse move or scroll events.
Limiting the number of events is not necessary anymore, because
processing them is now fast enough.