ui-client: update NPM dependencies
Updates npm dependencies in ui-client to be able to build with newer
node versions. Node 12 isn't available for aarch64 (Apple M1) and building
with a newer local node version would fail because of libscss being
outdated and not building. While updating webpack the the extract-text
plugin was replaces with mini-css-extract because extract-text is
deprecated.