New JavaScript function Modelix.openInProjector(nodeRef) to open a node in projector
It sends a POST request to an endpoint inside MPS that then opens the node
and switches MPS into a mode where toolbars, menus, status bars, tool windows, ...
are hidden. Only the editor for the node is visible. It looks exactly like
the image editor, but all the user input and image update handling is
done by projector.

The only disadvantage is that an MPS instance can't be shared between users,
but this causes a lot of problems anyway.