Sort modules into chunks
Dependencies must be generated first, but we want to generate in as few
steps as possible. Independent modules can be generated together.
The maximum number of cunks is equals to the longest path in the
dependency graph.

The generator could compute these chunks at runtime, but they are
expected to be definied in the ant script.