Start a new MPS instance after a change in the workspace definition
A version hash is assigned to each revision of a workspace definition.
This hash is then part of the URL that opens MPS.
Users can still continue working with the old version, but if they click
on the link in the workspace editor, it redirects to a new MPS instance.