Ignore modules that are not part of a project
If there is a .mps/modules.xml load only those modules that are listed
in this file and ignore all other files in that folder.

A project folder may contain files that were remove from the project,
but weren't deleted from the file system. They are not maintained
anymore and very likely broken and make the build fail.