The mps-tooltip plugin contains the pluginx.xml twice
Once in META-INF folder of the plugin directory and a second time in one
of the jars. The path to the jar was used in the generator script.
META-INF/plugin.xml is now ignored if the jar is inside a plugin
directory. Also an exeception is thrown if a plugin ID is found twice.